+++ src/app/parsers/CSVParser.ts
@@ -29,6 +29,8 @@ export class CSVParser extends InfinityParser {
           value = new Date(value);
         } else if (c.type === 'timestamp_epoch') {
           value = new Date(parseInt(value, 10));
+        } else if (c.type === 'timestamp_epoch_s') {
+          value = new Date(parseInt(value, 10) * 1000);
         } else if (c.type === 'number') {
           value = value === '' ? null : +value;
         }
@@ -55,6 +57,8 @@ export class CSVParser extends InfinityParser {
             timestamp = new Date(get(r, FirstTimeColumn.selector)).getTime();
           } else if (FirstTimeColumn.type === 'timestamp_epoch') {
             timestamp = new Date(parseInt(get(r, FirstTimeColumn.selector), 10)).getTime();
+          } else if (FirstTimeColumn.type === 'timestamp_epoch_s') {
+            timestamp = new Date(parseInt(get(r, FirstTimeColumn.selector), 10) * 1000).getTime();
           }
         }
         let metric = get(r, metricColumn.selector);