+++ src/app/parsers/JSONParser.ts
@@ -42,6 +42,8 @@ export class JSONParser extends InfinityParser {
           value = new Date(value + '');
         } else if (c.type === 'timestamp_epoch') {
           value = new Date(parseInt(value, 10));
+        } else if (c.type === 'timestamp_epoch_s') {
+          value = new Date(parseInt(value, 10) * 1000);
         } else if (c.type === 'number') {
           value = value === '' ? null : +value;
         }
@@ -68,6 +70,8 @@ export class JSONParser extends InfinityParser {
             timestamp = new Date(get(r, FirstTimeColumn.selector) + '').getTime();
           } else if (FirstTimeColumn.type === 'timestamp_epoch') {
             timestamp = new Date(parseInt(get(r, FirstTimeColumn.selector), 10)).getTime();
+          } else if (FirstTimeColumn.type === 'timestamp_epoch_s') {
+            timestamp = new Date(parseInt(get(r, FirstTimeColumn.selector), 10) * 1000).getTime();
           }
         }
         let metric = toNumber(get(r, metricColumn.selector));