+++ pkg/plugin.go
@@ -1,0 +1,72 @@
+package main
+
+import (
+	"context"
+	"encoding/json"
+	"net/http"
+
+	"github.com/grafana/grafana-plugin-sdk-go/backend"
+	"github.com/grafana/grafana-plugin-sdk-go/backend/datasource"
+	"github.com/grafana/grafana-plugin-sdk-go/backend/instancemgmt"
+	"github.com/grafana/grafana-plugin-sdk-go/data"
+)
+
+type queryModel struct{}
+
+func newDatasource() datasource.ServeOpts {
+	im := datasource.NewInstanceManager(newDataSourceInstance)
+	ds := &InfinityDatasource{
+		im: im,
+	}
+	return datasource.ServeOpts{
+		QueryDataHandler:   ds,
+		CheckHealthHandler: ds,
+	}
+}
+
+// InfinityDatasource is an example datasource used to scaffold
+type InfinityDatasource struct {
+	im instancemgmt.InstanceManager
+}
+
+// CheckHealth handles health checks sent from Grafana to the plugin.
+func (ds *InfinityDatasource) CheckHealth(ctx context.Context, req *backend.CheckHealthRequest) (*backend.CheckHealthResult, error) {
+	return &backend.CheckHealthResult{
+		Status:  backend.HealthStatusOk,
+		Message: "Data source is working",
+	}, nil
+}
+
+// QueryData handles multiple queries and returns multiple responses.
+func (ds *InfinityDatasource) QueryData(ctx context.Context, req *backend.QueryDataRequest) (*backend.QueryDataResponse, error) {
+	response := backend.NewQueryDataResponse()
+	for _, q := range req.Queries {
+		res := ds.query(ctx, q)
+		response.Responses[q.RefID] = res
+	}
+	return response, nil
+}
+
+func (ds *InfinityDatasource) query(ctx context.Context, query backend.DataQuery) (response backend.DataResponse) {
+	var qm queryModel
+	response.Error = json.Unmarshal(query.JSON, &qm)
+	if response.Error != nil {
+		return response
+	}
+	frame := data.NewFrame("response")
+	response.Frames = append(response.Frames, frame)
+
+	return response
+}
+
+type instanceSettings struct {
+	httpClient *http.Client
+}
+
+func (is *instanceSettings) Dispose() {}
+
+func newDataSourceInstance(setting backend.DataSourceInstanceSettings) (instancemgmt.Instance, error) {
+	return &instanceSettings{
+		httpClient: &http.Client{},
+	}, nil
+}