+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Change history of the project. All the feature updates, bug fixes, breaking chan
 
 - Plugin signed
 - JSONPath support for JSON root selector
+- Support for Unix/epoch seconds format
 
 ## [ 0.5.0 ]